reconstruction via trilinear merge

+++ src/chad/tsdf_map.cpp
@@ -10,7 +10,7 @@
 // TODO: another tree (addition to tsdf and weights) for ESDF with low res
 
 namespace chad {
-    TSDFMap::TSDFMap(float sdf_res, float sdf_trunc, float submap_xyz_threshhold, float submap_cor_threshhold):
+    TSDFMap::TSDFMap(double sdf_res, double sdf_trunc, double submap_xyz_threshhold, double submap_cor_threshhold):
         _sdf_res(sdf_res),
         _sdf_trunc(sdf_trunc),
         _dag_p(std::make_unique<detail::dag::Storage>()),
@@ -20,11 +20,11 @@ namespace chad {
     }
     void TSDFMap::clear() {
         _dag_p->clear();
-        // TODO: clear map optimizer
-        throw std::logic_error("Function not yet implemented: chad::TSDFMap::rebuild_hashes()");
+        // TODO: clear optimizer
+        throw std::logic_error("Function not yet implemented: chad::TSDFMap::clear()");
     }
     void TSDFMap::release_hashes() {
-        throw std::logic_error("Function not yet implemented: chad::TSDFMap::rebuild_hashes()");
+        _dag_p->release_hashes();
     }
     void TSDFMap::rebuild_hashes() {
         throw std::logic_error("Function not yet implemented: chad::TSDFMap::rebuild_hashes()");
@@ -66,25 +66,53 @@ namespace chad {
         using namespace chad::detail;
         // finalize current active submap if needed
         _map_optimizer_p->finalize();
+        if (_map_optimizer_p->_submaps.empty()) {
+            CHAD_MESSAGE("there is nothing to reconstruct!");
+            return;
+        }
+
+        // hashmaps are no longer needed at this point
+        release_hashes();
 
         // make sure the folder is clean
         if (clean_first) std::filesystem::remove_all(foldername);
         std::filesystem::create_directory(foldername);
 
-        // reconstruct every submap
-        for (std::uint32_t i = 0; i < _map_optimizer_p->_submaps.size(); i++) {
-            auto beg = std::chrono::steady_clock::now();
+        // create an octree that will be used for merging
+        map::Octree2<17, 2> octree_main;
+        Pose pose_main_orig = _map_optimizer_p->_submaps[0]._pose_avg;
+        Pose pose_main_real = _map_optimizer_p->posegraph_get(0);
+        glm::aligned_dquat main_rotation = glm::inverse(pose_main_real._rotation);
+        glm::aligned_dvec3 main_translation = pose_main_orig._position - pose_main_real._position;
 
-            // simply append index to the filename
-            const map::Submap& submap = _map_optimizer_p->_submaps[i];
-            std::string filename = fmt::format("{}/submap_{}.ply", foldername, i);
+        // merge other octrees into the base one
+        for (std::size_t submap_i = 0; submap_i < _map_optimizer_p->_submaps.size(); submap_i++) {
+            auto timestamp = std::chrono::steady_clock::now();
 
-            // reconstruct 3D mesh from the merged octree
-            reconstruction::reconstruct(filename, *_dag_p, submap._roots, _sdf_res, _sdf_trunc);
-            // reconstruction::reconstruct_points(filename, *_dag_p, submap._roots, _sdf_res, _sdf_trunc);
+            // grab the right submap and create an octree from its DAG tree
+            const map::Submap& submap_other = _map_optimizer_p->_submaps[submap_i];
+            map::Octree2<17, 2> octree_other{ *_dag_p, submap_other._roots, _sdf_trunc };
+            Pose pose_other_orig = submap_other._pose_avg;
+            Pose pose_other_real = _map_optimizer_p->posegraph_get(submap_i);
 
-            MEASURE_TIME(beg, fmt::format(">> Reconstructing submap \"{}\"", filename));
+            // calculate the transform from coordinate frame octree_other (B) to octree_main (A)
+            glm::aligned_dmat4x4 transform_b_to_a = glm::identity<glm::dmat4x4>();
+            transform_b_to_a = glm::translate(transform_b_to_a, pose_other_real._position + main_translation); // translate to coordinate frame A
+            transform_b_to_a = transform_b_to_a * glm::mat4_cast(pose_other_real._rotation * main_rotation); // rotate around origin
+            transform_b_to_a = glm::translate(transform_b_to_a, -pose_other_orig._position); // translate to origin
+
+            // merge octree_other into the main octree using the transform between their coordinate frames
+            octree_main.merge(octree_other, transform_b_to_a, _sdf_res);
+
+            double progress = static_cast<double>(submap_i + 1) / static_cast<double>(_map_optimizer_p->_submaps.size());
+            MEASURE_TIME(timestamp, fmt::format(">> Reconstructing submap [merge] {:6.2f}%", progress * 100.0));
         }
+
+        // create mesh and write it to disk
+        auto beg = std::chrono::steady_clock::now();
+        std::string filename = fmt::format("{}/chunk_{}.ply", foldername, 0);
+        reconstruction::Ply ply{ filename, octree_main, _sdf_res, true };
+        MEASURE_TIME(beg, fmt::format(">> Reconstructing \"{}\"", filename));
     }
     void TSDFMap::insert_internal(const std::uint8_t* data_p, std::size_t data_bytes, PointFlags data_flags, std::array<double, 3> position, std::array<double, 3> rotation) {
         auto beg = std::chrono::steady_clock::now();
